Squashed commit of the following:
commit f878a00ddd048f9a79fb0ed593539a7a5d21b773
Merge: 09aa7fd aa27730
Author: kjw402540 <kjw402540@gmail.com>
Date:   Fri Dec 5 15:10:12 2025 +0900

    Merge branch 'jiwoo_1203' of https://github.com/kjw402540/castaway into jiwoo_1203

commit 09aa7fdeb297876ebb0a3ce16aa864aabec2e7fe
Author: kjw402540 <kjw402540@gmail.com>
Date:   Fri Dec 5 15:10:08 2025 +0900

    리포트테스트코드

commit 83857ca6a9794d839106230ac7f0d964b0c01be8
Author: kjw402540 <kjw402540@gmail.com>
Date:   Wed Dec 3 15:37:16 2025 +0900

    오브제생성

    bgm수정

    오브젝트생성

    오브제2

    프리즈마

    Object -> object

    objects -> object

    오브제3

    오브제4

    오브제5

commit aa277302dbe05fa28a4f042c09ec3fe2534b708f
Author: kjw402540 <kjw402540@gmail.com>
Date:   Fri Dec 5 10:48:56 2025 +0900

    오브제6

commit a136df1f01fbc5680916cadde978a29a65b93c07
Author: kjw402540 <kjw402540@gmail.com>
Date:   Fri Dec 5 10:16:06 2025 +0900

    오브제5

commit 55f96cd3e385831429dfccb76e7e9ba42afedb8b
Author: kjw402540 <kjw402540@gmail.com>
Date:   Thu Dec 4 15:04:38 2025 +0900

    오브제4

commit d5a1898b4e3f540486e946e388138e54c71c14d4
Author: kjw402540 <kjw402540@gmail.com>
Date:   Thu Dec 4 14:51:17 2025 +0900

    오브제3

commit 71dcfbebd9281c6b1dff3abecbf441ff4aa36d24
Author: kjw402540 <kjw402540@gmail.com>
Date:   Thu Dec 4 11:44:45 2025 +0900

    objects -> object

commit 36a19beadcab210236d28f0ae4b456e13bcfc5a6
Author: kjw402540 <kjw402540@gmail.com>
Date:   Thu Dec 4 11:40:09 2025 +0900

    프리즈마

    Object -> object

commit dea8dd0daa96c188338428e54f9eebf8b8d6c1ce
Author: kjw402540 <kjw402540@gmail.com>
Date:   Thu Dec 4 11:31:00 2025 +0900

    오브제2

commit d59dbbe742e8e7ea169f524ae529d1162a60c7f5
Author: kjw402540 <kjw402540@gmail.com>
Date:   Thu Dec 4 11:05:26 2025 +0900

    오브젝트생성

commit f742318b0987b0eb6d7fb30ce0fe1d14aa782004
Author: kjw402540 <kjw402540@gmail.com>
Date:   Wed Dec 3 15:37:16 2025 +0900

    bgm수정

+++ apps/api/src/controllers/objectController.js
@@ -7,7 +7,9 @@ import * as diaryService from "../services/diaryService.js";
 ----------------------------------------- */
 export const getAll = async (req, res, next) => {
   try {
-    const userId = 1;
+    // ⚠️ 미들웨어(authMiddleware.js)가 'id'로 저장하므로 .id로 접근해야 함!
+    const userId = req.user.id; 
+    
     const list = await objectService.getAll(userId);
     res.json(list);
   } catch (err) {
@@ -16,12 +18,13 @@ export const getAll = async (req, res, next) => {
 };
 
 /* ----------------------------------------
-   날짜별 오브제 (GET /object/:date)
+   날짜별 오브제
 ----------------------------------------- */
 export const getByDate = async (req, res, next) => {
   try {
-    const userId = 1;
+    const userId = req.user.id; // 여기도 수정
     const { date } = req.params;
+    
     const object = await objectService.getByDate(userId, date);
     res.json(object);
   } catch (err) {
@@ -43,18 +46,22 @@ export const getById = async (req, res, next) => {
 };
 
 /* ----------------------------------------
-   오브제 삭제 = 일기 세트 삭제
+   오브제 삭제
 ----------------------------------------- */
 export const remove = async (req, res, next) => {
   try {
     const { id } = req.params;
+    const userId = req.user.id; // 삭제 권한 확인용 (필요시 사용)
 
     const obj = await objectService.getById(id);
     if (!obj) return res.status(404).json({ error: "Not found" });
 
-    const diaryId = obj.diary_id;
+    // (옵션) 본인 것인지 확인
+    if (obj.user_id !== userId) {
+        return res.status(403).json({ error: "권한이 없습니다." });
+    }
 
-    // ✔ diaryService.removeById 사용
+    const diaryId = obj.diary_id;
     const removed = await diaryService.removeById(diaryId);
 
     res.json({ ok: true, removed });
@@ -64,7 +71,7 @@ export const remove = async (req, res, next) => {
 };
 
 /* ----------------------------------------
-   오브제 배치 (위치 저장 placeholder)
+   오브제 배치
 ----------------------------------------- */
 export const place = async (req, res, next) => {
   try {
